Split up SearchResults window logic into separate files
