+++ stockflux-launcher/src/free-text-search/search-result/SearchResults.populater.js
@@ -1,0 +1,6 @@
+import ReactDOMServer from 'react-dom/server';
+
+export default (html, resultsWindow) => {
+  const markup = ReactDOMServer.renderToStaticMarkup(html);
+  resultsWindow.getWebWindow().document.getElementById('results-container').innerHTML = markup;
+};
